pyproject.toml
[tool.poetry]
name = "f5mobile-dashboard"
version = "0.1.0"
description = ""
authors = ["tf_01"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
requests = "^2.32.3"
httpx = "^0.27.0"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
